requirements.txt
omegaconf
psutil
matplotlib
annoy
hnswlib
scann
faiss-cpu
sklearn